Improved Path joins
